Learned about variables and functions
I now have a basic understanding of functions and variables in PICO-8.

`_init()`, `_update()` and `_draw()` are special functions that form the
game loop.

I managed to use `_update()` and `_draw()` to create a ❤️  that moves
across the screen.

Conditions were not introduced but I decided to see if they are similar
to Ruby. I don't think Ruby requires me to use "then" but Lua seems to.