Fix ball deflecting incorrectly
When the ball hits two bricks it would earlier deflect off both. This
caused an incorrect deflection.

This commit updates the deflection algorithm to deflect off of only one
brick if it hits the edges of two. Since it is a hit, both bricks will
disappear but the deflection will be off only the first one.